[SAFE-399] accessToken input popup added + link of the form added in the request return
