+++ routes/download/index.ts
@@ -8,10 +8,13 @@ import fs from 'fs';
 import fileBuilder from "../../utils/files/fileBuilder";
 import koboBuilder from "../../utils/files/koboBuilder";
 import request from "request"
+import bodyParser from "body-parser";
+import exportForms from "../../utils/exportForms";
 
 /* CSV or xlsx export of records attached to a form.
 */
 const router = express.Router();
+
 router.get('/form/records/:id', async (req, res) => {
     const ability: AppAbility = req.context.user.ability;
     const filters = Form.accessibleBy(ability, 'read').where({_id: req.params.id}).getFilter();
@@ -113,134 +116,6 @@ router.get('/file/:form/:blob', async (req, res) => {
             console.log('file deleted');
         });
     });
-});
-
-/* CSV or xlsx export of a form.
-*/
-router.get('/form/kobo/:id', async (req, res) => {
-    const ability: AppAbility = req.context.user.ability;
-    const filters = Form.accessibleBy(ability, 'read').where({_id: req.params.id}).getFilter();
-    const form = await Form.findOne(filters);
-    let buffer;
-    let uid1;
-    let uid2;
-    if (form) {
-
-        // CREATE EXCEL FILE
-
-        buffer = await koboBuilder(res, form);
-        console.log('*** buffer ***');
-        console.log(buffer);
-
-        // IMPORT EXCEL FILE
-
-        const options = {
-            'method': 'POST',
-            'url': 'https://kobo.humanitarianresponse.info/api/v2/imports/?format=json',
-            'headers': {
-                'Authorization':`Token ${process.env.TOKEN_KOBO}`,
-                'Accept': 'application/json',
-                // 'Cookie': 'csrftoken=tEJ8gIf31yRCPt2nWt62UnZH7yZBKuiv4elnE8YKu2OZTz3p0sAYuY5hc4qU3qPy'
-            },
-            formData: {
-                'file': {
-                    'value': buffer,
-                    // 'value': fs.createReadStream('/Users/martin/Desktop/Stage_ReliefApp/dev/divers/test_import.xlsx'),
-                    'options': {
-                        'filename': 'test_import.xlsx',
-                        'contentType': null
-                    }
-                },
-                'library': 'false',
-                'name': 'form_'+Date.now().toString(),
-            }
-        };
-        await request(options, function (error, response) {
-            if (error) throw new Error(error);
-            console.log('### FILE SEND ###');
-            console.log(response.body);
-            const body = JSON.parse(response.body.toString());
-            console.log(body);
-            console.log(body['uid']);
-            console.log(body.uid);
-            uid1 = body.uid;
-            console.log(uid1);
-            console.log(`https://kobo.humanitarianresponse.info/api/v2/imports/${uid1}/?format=json`);
-
-            // GET UID OF THE NEW FORM
-
-            const options = {
-                'method': 'GET',
-                'url': `https://kobo.humanitarianresponse.info/api/v2/imports/${uid1}/?format=json`,
-                'headers': {
-                    'Authorization':`Token ${process.env.TOKEN_KOBO}`,
-                    'Accept': 'application/json',
-                    // 'Cookie': 'csrftoken=tEJ8gIf31yRCPt2nWt62UnZH7yZBKuiv4elnE8YKu2OZTz3p0sAYuY5hc4qU3qPy'
-                }
-            };
-            request(options, function (error, response) {
-                if (error) throw new Error(error);
-                console.log('@@@ GET UID @@@');
-                console.log(response.body);
-                const body = JSON.parse(response.body.toString());
-                uid2 = body.messages.created[0].uid;
-                console.log('*** uid2 ***');
-                console.log(uid2);
-
-                // DEPLOY FORM
-
-                const options = {
-                    'method': 'POST',
-                    'url': `https://kobo.humanitarianresponse.info/api/v2/assets/${uid2}/deployment/?format=json`,
-                    'headers': {
-                        'Authorization': `Token ${process.env.TOKEN_KOBO}`,
-                        // 'Cookie': 'csrftoken=tEJ8gIf31yRCPt2nWt62UnZH7yZBKuiv4elnE8YKu2OZTz3p0sAYuY5hc4qU3qPy'
-                    },
-                    formData: {
-                        'active': 'true'
-                    }
-                };
-                request(options, function (error, response) {
-                    if (error) throw new Error(error);
-                    console.log(response.body);
-                    const body = JSON.parse(response.body.toString());
-                    const url = body.asset.deployment__links.url;
-                    // TODO:
-                    // res.send(url)
-                    // manage the return of this request (before the frontend was waiting for a file, that not the case anymore)
-                    // display the url in the frontend
-                    // and store it somewhere
-                });
-
-            });
-
-            // setTimeout(() => {
-            //     // GET UID OF THE NEW FORM
-            //
-            //     const options = {
-            //         'method': 'GET',
-            //         'url': `https://kobo.humanitarianresponse.info/api/v2/imports/${uid1}/?format=json`,
-            //         'headers': {
-            //             'Authorization':`Token ${process.env.TOKEN_KOBO}`,
-            //             'Accept': 'application/json',
-            //             // 'Cookie': 'csrftoken=tEJ8gIf31yRCPt2nWt62UnZH7yZBKuiv4elnE8YKu2OZTz3p0sAYuY5hc4qU3qPy'
-            //         }
-            //     };
-            //     request(options, function (error, response) {
-            //         if (error) throw new Error(error);
-            //         console.log('@@@ GET UID @@@');
-            //         console.log(response.body);
-            //         uid2 = response.body.messages.created[0].uid;
-            //         console.log('*** uid2 ***');
-            //         console.log(uid2);
-            //     });
-            // }, 5000);
-
-        });
-
-    } else {
-        res.status(404).send(errors.dataNotFound);
-    }
 });
 
 export default router;
